+++ appWeb/appWeb/build.gradle.kts
@@ -1,3 +1,3 @@
 plugins {
-    id("web-browser-executable-plugin")
+    id("web-executable-plugin")
 }
