+++ gradle-plugins/build.gradle.kts
@@ -19,15 +19,14 @@ gradlePlugin {
         id = "build-config-plugin"
         implementationClass = "com.gchristov.thecodinglove.gradleplugins.BuildConfigPlugin"
     }
-    plugins.register("web-browser-executable-plugin") {
-        id = "web-browser-executable-plugin"
-        implementationClass =
-            "com.gchristov.thecodinglove.gradleplugins.web.WebBrowserExecutablePlugin"
+    plugins.register("web-executable-plugin") {
+        id = "web-executable-plugin"
+        implementationClass = "com.gchristov.thecodinglove.gradleplugins.web.WebExecutablePlugin"
     }
-    plugins.register("backend-node-executable-plugin") {
-        id = "backend-node-executable-plugin"
+    plugins.register("backend-executable-plugin") {
+        id = "backend-executable-plugin"
         implementationClass =
-            "com.gchristov.thecodinglove.gradleplugins.backend.BackendNodeExecutablePlugin"
+            "com.gchristov.thecodinglove.gradleplugins.backend.BackendExecutablePlugin"
     }
 }
