+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/web/WebExecutablePlugin.kt
@@ -1,18 +1,21 @@
-package com.gchristov.thecodinglove.gradleplugins.backend
+package com.gchristov.thecodinglove.gradleplugins.web
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
 @Suppress("unused")
-class BackendNodeExecutablePlugin : Plugin<Project> {
+class WebExecutablePlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.plugins.apply("org.jetbrains.kotlin.multiplatform")
-        target.plugins.apply("dev.petuska.npm.publish")
         target.extensions.configure(KotlinMultiplatformExtension::class.java) {
             js(IR) {
-                binaries.library()
-                nodejs()
+                binaries.executable()
+                browser {
+                    commonWebpackConfig {
+                        cssSupport.enabled = true
+                    }
+                }
             }
         }
     }
